move role-authorization to module and create gen.js for this module
